feat(ai-pow-miner): isolate CUDA miner dependencies
Extract node-miner work and reward arguments into a shared node-gated CLI module, then add an explicitly excluded CUDA miner package pinned to cuda-oxide 0810ba95. The CUDA host session owns a validated sm80-or-newer context and dedicated stream, and the CUDA entrypoint has no CPU fallback. Normal workspace builds remain free of CUDA dependency resolution.
