+++ crates/ai-pow-miner-cuda/rust-toolchain.toml
@@ -1,0 +1,3 @@
+[toolchain]
+channel = "nightly-2026-04-03"
+components = ["rust-src", "rustc-dev"]
